+++ code/cell.modules/pkg.sys/fs/src/web/types.ts
@@ -8,5 +8,5 @@ export { Fs } from '@platform/cell.types/lib/types.Fs';
  * @local
  */
 export * from './FsDriver.IndexedDb/types';
-export * from './ui/PathList/types';
+export * from './ui/FsPathList/types';
 export * from './ui/FsCard/types';
