+++ code/cell.modules/pkg.sys/fs/src/web/ui/FsPathList/FsPathList.tsx
@@ -1,6 +1,6 @@
 import { DEFAULT, FC, THEMES } from './common';
 import { usePathListState } from './logic/usePathListState';
-import { PathListStateful as Stateful } from './PathList.Stateful';
+import { FsPathListStateful as Stateful } from './FsPathList.Stateful';
 import { PathListProps } from './types';
 import { PathList as View } from './ui/PathList';
 import { wrangle } from './ui/wrangle';
@@ -15,8 +15,8 @@ type Fields = {
   THEMES: typeof THEMES;
   DEFAULT: typeof DEFAULT;
 };
-export const PathList = FC.decorate<PathListProps, Fields>(
+export const FsPathList = FC.decorate<PathListProps, Fields>(
   View,
   { Stateful, useState: usePathListState, wrangle, THEMES, DEFAULT },
-  { displayName: 'PathList' },
+  { displayName: 'Fs.PathList' },
 );
