+++ code/cell.modules/pkg.sys/fs/src/web/ui/FsPathList/index.ts
@@ -1,0 +1,3 @@
+export * from './FsPathList';
+export * from './FsPathList.Stateful';
+export { PathListProps, PathListStatefulProps } from './types';
