+++ code/cell.modules/pkg.sys/fs/src/web/ui/FsPathList/ui/renderers.tsx
