--- code/cell.modules/pkg.sys/fs/src/web/ui/PathList/index.ts
@@ -1,3 +1,0 @@
-export * from './PathList';
-export * from './PathList.Stateful';
-export { PathListProps, PathListStatefulProps } from './types';
